Move Artifice registry functions to top-level
