+++ src/main/java/com/swordglowsblue/artifice/api/Artifice.kt
@@ -45,72 +45,4 @@ object Artifice {
         ArtificeImpl.LOGGER.warn("Using deprecated Artifice#registerData! Please use registerDataPack! Issues may occur!")
         return ArtificeImpl.registerSafely(ArtificeRegistry.DATA, id, ArtificeResourcePack.ofData(register))
     }
-
-    /**
-     * Register a new server-side resource pack, creating resources with the given callback.
-     *
-     * @param id       The pack ID.
-     * @param register A callback which will be passed a [ServerResourcePackBuilder].
-     * @return The registered pack.
-     */
-    @JvmStatic
-    fun registerDataPack(id: Identifier, register: Processor<ServerResourcePackBuilder>) {
-        ArtificeImpl.registerSafely(
-            ArtificeRegistry.DATA,
-            id,
-            DynamicResourcePackFactory(ResourceType.SERVER_DATA, id, register)
-        )
-    }
-
-    /**
-     * Register a new client-side resource pack.
-     *
-     * @param id   The pack ID.
-     * @param pack The pack to register.
-     * @return The registered pack.
-     */
-    @Environment(EnvType.CLIENT)
-    fun registerAssets(id: String, pack: ArtificeResourcePack): ArtificeResourcePack {
-        return registerAssets(Identifier(id), pack)
-    }
-
-    /**
-     * Register a new server-side resource pack.
-     *
-     * @param id   The pack ID.
-     * @param pack The pack to register.
-     * @return The registered pack.
-     */
-    fun registerData(id: String, pack: ArtificeResourcePack): ArtificeResourcePack {
-        return registerData(Identifier(id), pack)
-    }
-
-    /**
-     * Register a new client-side resource pack.
-     *
-     * @param id   The pack ID.
-     * @param pack The pack to register.
-     * @return The registered pack.
-     * @see ArtificeResourcePack.ofAssets
-     */
-    @JvmStatic
-    @Environment(EnvType.CLIENT)
-    fun registerAssets(id: Identifier, pack: ArtificeResourcePack): ArtificeResourcePack {
-        require(pack.type == ResourceType.CLIENT_RESOURCES) { "Cannot register a server-side pack as assets" }
-        return ArtificeImpl.registerSafely(ArtificeRegistry.ASSETS, id, pack)
-    }
-
-    /**
-     * Register a new server-side resource pack.
-     *
-     * @param id   The pack ID.
-     * @param pack The pack to register.
-     * @return The registered pack.
-     * @see ArtificeResourcePack.ofData
-     */
-    @JvmStatic
-    fun registerData(id: Identifier, pack: ArtificeResourcePack): ArtificeResourcePack {
-        require(pack.type == ResourceType.SERVER_DATA) { "Cannot register a client-side pack as data" }
-        return ArtificeImpl.registerSafely(ArtificeRegistry.DATA, id, pack)
-    }
 }
